Revert "allow short positions in PositionTable"
This reverts commit 924f2e79076f5f8a76f35630dfd4ff9463954325.

+++ system/src/de/uka/ilkd/key/pp/PositionTable.java
@@ -39,10 +39,10 @@ import de.uka.ilkd.key.logic.PosInTerm;
 public class PositionTable {
 
     // the start positions of the direct subterms (or parts of sequent, etc.)
-    protected short[] startPos;
+    protected byte[] startPos;
 
     //the end positions of the direct subterms (or parts of sequent, etc.)   
-    protected short[] endPos;
+    protected byte[] endPos;
 
     // the PositionTables for the direct subterms (or parts of sequent, etc.)
     protected PositionTable[] child;    
@@ -55,7 +55,7 @@ public class PositionTable {
 
     // the number of rows in the above arrays. Equals the number of direct 
     // subterms (or parts of sequent, etc.)
-    private final short rows;
+    private final byte rows;
 
     /**
      * creates a new PositionTable with the number of subterms (or number of
@@ -65,14 +65,14 @@ public class PositionTable {
      * position information is represented by the constructed object.
      */
     public PositionTable(int rows) {
-        if (rows < Short.MIN_VALUE || rows > Short.MAX_VALUE)
+        if (rows < Byte.MIN_VALUE || rows > Byte.MAX_VALUE)
             throw new ArithmeticException("The number of subterms "+rows+" exceeds the designed range.");
         
-	this.rows=(short)rows;
-	startPos=new short[this.rows];
-	endPos=new short[this.rows];
+	this.rows=(byte)rows;
+	startPos=new byte[this.rows];
+	endPos=new byte[this.rows];
 	child=new PositionTable[this.rows];
-	for (short i=0; i<this.rows; i++) {
+	for (byte i=0; i<this.rows; i++) {
 	    startPos[i]=-1;
 	    endPos[i]=-1;
 	    child[i]=null;
@@ -206,9 +206,9 @@ public class PositionTable {
      * @param child PositionTable for the sub-element from start to end
      */
     public void setEnd(int end, PositionTable child) {
-        if (end < Short.MIN_VALUE || Short.MAX_VALUE < end)
+        if (end < Byte.MIN_VALUE || Byte.MAX_VALUE < end)
             throw new ArithmeticException("Position "+end+" exceeds the design limits.");
-	endPos[currentEndEntry]=(short)end;
+	endPos[currentEndEntry]=(byte)end;
 	this.child[currentEndEntry]=child;
 	currentEndEntry++;
     }
@@ -219,9 +219,9 @@ public class PositionTable {
      * @param start char position that starts a sub-element
      */
     public void setStart(int start) {
-        if (start < Short.MIN_VALUE || Short.MAX_VALUE < start)
+        if (start < Byte.MIN_VALUE || Byte.MAX_VALUE < start)
             throw new ArithmeticException("Position "+start+" exceeds the design limits.");
-	startPos[currentStartEntry]=(short)start;
+	startPos[currentStartEntry]=(byte)start;
 	currentStartEntry++;
     }
